backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-JWT-Extended==4.6.0
scikit-learn==1.5.1
pymongo==4.8.0
bcrypt==4.2.0
python-dotenv==1.0.1
